\name{expressions....}
\alias{expressions....}
\title{Extract the expressions from a dots object.}
\usage{
  expressions....(x)
}
\arguments{
  \item{x}{A dots object (see \code{\link{dots}}}
}
\value{
  A named list of expressions.
}
\description{
  Extract the expressions from a dots object.
}
\seealso{
  dots unpack
}

