\name{expressions_and_pointers}
\alias{expressions_and_pointers}
\title{Passed a list of dots arguments, returns their expressions and, in
names, their machine pointers}
\usage{
  expressions_and_pointers(...)
}
\arguments{
  \item{...}{A varying number of arguments.}
}
\value{
  A vector of integers containing pointer values, one per
  argument.
}
\description{
  Passed a list of dots arguments, returns their
  expressions and, in names, their machine pointers
}
\author{
  Peter Meilstrup
}

