\name{getPointer}
\alias{getPointer}
\title{Return the raw pointer of each SEXP given as argument.}
\usage{
  getPointer(...)
}
\arguments{
  \item{...}{A varying number of arguments.}
}
\value{
  A vector of integers containing raw pointers.
}
\description{
  Print some information about the arguments, including
  hopefully pointers to sexps, for expression arguments.
}
\author{
  Peter Meilstrup
}

