% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/labelOutlier.R
\name{labelOutlier}
\alias{labelOutlier}
\title{labelOutlier}
\usage{
labelOutlier(x)
}
\arguments{
\item{x}{}
}
\value{
labeled x
}
\description{
Currently this won't be included in the package.
}
\examples{
an example
}
\keyword{internal}

