% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cache.R
\name{macro_cache_report}
\alias{macro_cache_report}
\title{Report on macro cache contents.}
\usage{
macro_cache_report()
}
\value{
list with entries "hits", "misses", and "size"
}
\description{
Report on macro cache contents.
}
\author{
Peter Meilstrup
}

