% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mply.R
\name{mply}
\alias{mply}
\title{Alternative to \code{\link{mapply}} with a cleaner calling convention.}
\usage{
mply(...)(...)
}
\arguments{
\item{...}{In the first set of parens, a function to apply, and extra
arguments to be applied to each call. In the second set of parens, arguments
to iterate over. Multiple arguments will be recycled up to the length of the
longest argument.}
}
\value{
A list.
}
\description{
\code{mply} applies a function to corresponding elements from any number
of vector arguments.
}
\author{
Peter Meilstrup
}

