% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getpromise.R
\name{arg_dots}
\alias{arg_dots}
\alias{get_dots}
\title{Fetch promises associated with named arguments.}
\usage{
arg_dots(...)

get_dots(names, envir = arg_env(names, environment()))
}
\arguments{
\item{...}{Variable names (unevaluated). Arguments may be named; these names
determine the names on the dots list (and not the variable names)}

\item{names}{A character vector or list of names.}

\item{envir}{The environment to look for the argument names in. By default
looks in the lexical environment of the \code{name} argument.}
}
\value{
a \code{\link{dots}} object containing the promises that are bound to
those variables in the calling environment.
}
\description{
Fetch promises associated with named arguments.

\code{arg_get} fetches arguments from a named list.
}
\note{
The tags on the dots object are determined by argument names;
variable names are discarded.
}
\author{
Peter Meilstrup
}
\seealso{
dots get_dots
}

