\name{is.missing}
\alias{is.missing}
\title{Check if list members are equal to the "missing value."}
\usage{
  is.missing(x)
}
\arguments{
  \item{x}{An object}
}
\value{
  A vector of boolean values.
}
\description{
  For \code{\dots} objects as made by \code{\link{dots}},
  performs this check without forcing evaluation.
}
\author{
  Peter Meilstrup
}
\seealso{
  dots dots_missing missing_value
}

